remove legacy token list
