+++ web/src/routes/+page.svelte
@@ -8,7 +8,7 @@
 
 <div class="container h-full mx-auto flex justify-center items-center">
 	<div class="space-y-10 text-center flex flex-col items-center">
-		<Account />
+		<Account {promise} />
 		<Portfolio {promise} />
 	</div>
 </div>
